Replace SkipPackageSuppliers.default_ usage with Nullable
Instead of having a different enum value for stating that the value is
not specified and it should be taken from the configuration files, use
a Nullable. This has been suggested in
https://github.com/dlang/dub/pull/2903#discussion_r1582262457

The main reason for this change is going through the other places
where the configuration files were not honored (e.g. dub lint) (which
was the intent of the original PR). It would be a shorter change to
modify the current code to use the .default_ enum value when it was
appropriate but if I have to change the code I might as well go and
implement the less error-prone solution with Nullable.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>